Update path in GitHub deploy workflow
The deployment path in the GitHub workflow 'deploy.yml' file has been updated, changing from 'book' to 'connector-docs/book'. This change targets the documentation deployment process in the project's GitHub Pages workflow. It helps to direct the automatic deployment to the correct directory, facilitating better organization and access to the documentation for the connector tools.